requirements.txt
Flask==3.0.2
pymongo==4.6.1
python-dotenv==1.0.1
bcrypt==4.1.2
Flask-Session==0.6.0
google-generativeai==0.3.2
python-jose==3.3.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.1
gunicorn==21.2.0 